Workflow Changes | introduce a workflow that runs on PR to the main branch
