+++ requirements_ci.txt
@@ -1,0 +1,3 @@
+pylint
+black
+isort
